Repository/code/yahoo-finance-scraper/requirements.txt
elasticsearch
ruamel.yaml
yahoo-finance